move view requirements button to separate component
