+++ client/src/components/TestSuite/Requirements/styles.tsx
@@ -28,4 +28,8 @@ export default makeStyles()((theme: Theme) => ({
       margin: 0,
     },
   },
+  textButton: {
+    fontWeight: 'bolder',
+    cursor: 'pointer',
+  },
 }));
