+++ client/src/components/TestSuite/TestSuiteDetails/TestGroupListItem/NestedDescriptionPanel.tsx
@@ -1,26 +1,22 @@
 import React, { FC } from 'react';
 import Markdown from 'react-markdown';
 import remarkGfm from 'remark-gfm';
-import { enqueueSnackbar } from 'notistack';
-
 import {
   Accordion,
   AccordionDetails,
   AccordionSummary,
   Box,
   Divider,
-  Link,
   List,
   ListItem,
   ListItemText,
   Typography,
 } from '@mui/material';
 import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
-import { getSingleRequirement } from '~/api/RequirementsApi';
-import { Requirement, TestGroup } from '~/models/testSuiteModels';
+import { TestGroup } from '~/models/testSuiteModels';
 import useStyles from './styles';
 import { shouldShowRequirementsButton } from '~/components/TestSuite/TestSuiteUtilities';
-import RequirementsModal from '~/components/TestSuite/Requirements/RequirementsModal';
+import RequirementsModalButton from '~/components/TestSuite/Requirements/RequirementsModalButton';
 
 interface NestedDescriptionPanelProps {
   testGroup: TestGroup;
@@ -28,26 +24,8 @@ interface NestedDescriptionPanelProps {
 
 const NestedDescriptionPanel: FC<NestedDescriptionPanelProps> = ({ testGroup }) => {
   const { classes } = useStyles();
-  const [showRequirements, setShowRequirements] = React.useState(false);
-  const [requirements, setRequirements] = React.useState<Requirement[]>([]);
   const [descriptionMouseHover, setDescriptionMouseHover] = React.useState(false);
   const showRequirementsButton = shouldShowRequirementsButton(testGroup);
-
-  const showRequirementsClick = () => {
-    const requirementIds = testGroup.verifies_requirements;
-    if (requirementIds) {
-      Promise.all(requirementIds.map((requirementId) => getSingleRequirement(requirementId)))
-        .then((resolvedValues) => {
-          setRequirements(resolvedValues.filter((r) => !!r));
-          setShowRequirements(true);
-        })
-        .catch((e: Error) => {
-          enqueueSnackbar(`Error fetching specification requirements: ${e.message}`, {
-            variant: 'error',
-          });
-        });
-    }
-  };
 
   return (
     <Box py={1} className={classes.descriptionContainer}>
@@ -85,27 +63,10 @@ const NestedDescriptionPanel: FC<NestedDescriptionPanelProps> = ({ testGroup }) 
         >
           <Box className={`${classes.accordionDetail} ${classes.description}`}>
             <Markdown remarkPlugins={[remarkGfm]}>{testGroup.description as string}</Markdown>
-            {showRequirementsButton && (
-              <Box display="flex" justifyContent="end" minWidth="fit-content" pt={1}>
-                <Link
-                  color="secondary"
-                  className={classes.textButton}
-                  onClick={showRequirementsClick}
-                >
-                  View Specification Requirements
-                </Link>
-              </Box>
-            )}
+            {showRequirementsButton && <RequirementsModalButton runnable={testGroup} />}
           </Box>
         </AccordionDetails>
       </Accordion>
-      {requirements && showRequirementsButton && (
-        <RequirementsModal
-          requirements={requirements}
-          modalVisible={showRequirements}
-          hideModal={() => setShowRequirements(false)}
-        />
-      )}
     </Box>
   );
 };
